refactor(unimplemented message): rename component dir to camelCase
Fixes #388
