+++ app/renderer/ui/components/unimplementedMessage/style.scss
